bridge: drop packages if there is no default route on mesh table
Usally we will drop the packages on the mesh default due to the
unreachable default, in the case we accept default routes from
some peer this rule will be replaced. If this rule vanishes there
is no rule for default rejection of packages, thus they will be handled
by the default routing table, which should not happen.