maintenance: introduce maintenance mode
To be sure you do not harm user connection handled by the gateway, we
are now able to bring the host into maintenance mode. In that mode
the check-gateway script will no longer activate the gateway feature
of batman-adv.

TODO We need to bring down router advertisements to this host as well.